Task #5: Further refactoring performed. Changed comments, revised project configuration, reduced main.py
