+++ constants/limits.py
@@ -1,0 +1,3 @@
+LOGS_TO_CHOP = 3
+COBBLESTONE_TO_MINE = 20
+IRON_TO_MINE = 3
