+++ src/OvergroundActionShaper.py
@@ -48,56 +48,3 @@ class OvergroundActionShaper(gym.ActionWrapper, ABC):
     # action() is a method that must be overwritten so that step() could work correctly
     def action(self, action):
         return self.new_action_space[action]
-
-
-# This function gets the gym.spaces.Dict of actions and returns a numpy array of active actions during each step
-def normalize_actions(actions, batch_size, vertical_padding=7.5, horizontal_padding=5):
-    camera_actions = actions["camera"].squeeze()
-    attack_actions = actions["attack"].squeeze()
-    forward_actions = actions["forward"].squeeze()
-    sprint_actions = actions["sprint"].squeeze()
-    back_actions = actions["back"].squeeze()
-    left_actions = actions["left"].squeeze()
-    right_actions = actions["right"].squeeze()
-    jump_actions = actions["jump"].squeeze()
-
-    actions = np.zeros((batch_size,), dtype=np.int)
-
-    for i in range(len(camera_actions)):
-        # Moving camera has the highest priority
-        if camera_actions[i][0] < -horizontal_padding:
-            actions[i] = 7
-        elif camera_actions[i][0] > horizontal_padding:
-            actions[i] = 8
-        elif camera_actions[i][1] > vertical_padding:
-            actions[i] = 9
-        elif camera_actions[i][1] < -vertical_padding:
-            actions[i] = 10
-
-        # Then jump with/without moving forward
-        elif jump_actions[i] and forward_actions[i]:
-            if sprint_actions[i]:
-                actions[i] = 5
-            else:
-                actions[i] = 6
-
-        # Just move forward if there is no jumping action
-        elif forward_actions[i]:
-            actions[i] = 4
-
-        # Then other navigation actions
-        elif back_actions[i]:
-            actions[i] = 1
-        elif left_actions[i]:
-            actions[i] = 2
-        elif right_actions[i]:
-            actions[i] = 3
-
-        # Attacking has the lowest priority
-        elif attack_actions[i]:
-            actions[i] = 0
-        else:
-            # No reasonable mapping (will be ignored after applying a mask)
-            actions[i] = -1
-
-    return actions
